EnglistLearn/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
google-adk==1.18.0
pydantic>=2.10.6
python-dotenv>=1.0.1
aiofiles==23.2.1
sse-starlette==1.6.5